fair vis - refactored license into separate component
